// mobile/src/app/interfaces/banner.ts
export interface Banner {
  id: number;
  title: string;
  image_path: string;
  image_path_url: string;
  type: string;
  link_url: string;
  internal_category: string;
  internal_entity_id: number;
  internal_entity_name: string;
  status: number;
  status_label: string;
  created_at: number;
  updated_at: number;
}